Prepare `src/metabase` code for Hosted Bundle
