+++ enterprise/frontend/src/metabase-enterprise/whitelabel/components/ColorSettings/ColorSettings.unit.spec.tsx
@@ -2,7 +2,8 @@ import userEvent from "@testing-library/user-event";
 import Color from "color";
 
 import { render, screen } from "__support__/ui";
-import { color, colors } from "metabase/lib/colors/palette";
+import { colors } from "metabase/lib/colors/colors";
+import { color } from "metabase/lib/colors/palette";
 
 import { ColorSettings } from "./ColorSettings";
