+++ enterprise/frontend/src/metabase-enterprise/whitelabel/lib/whitelabel.js
@@ -1,4 +1,4 @@
-import { colors } from "metabase/lib/colors/palette";
+import { colors } from "metabase/lib/colors/colors";
 import MetabaseSettings from "metabase/lib/settings";
 
 export function updateColors() {
