+++ frontend/src/metabase/common/components/Alert/Alert.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from "react";
 
 import type { IconName } from "metabase/ui";
 
-import { AlertIcon, AlertRoot } from "./Alert.styled";
+import { AlertIcon, AlertRoot, CloseIcon } from "./Alert.styled";
 
 export type AlertVariant = "info" | "warning" | "error";
 
@@ -12,6 +12,7 @@ export interface AlertProps {
   hasBorder?: boolean;
   className?: string;
   variant?: AlertVariant;
+  onClose?: () => void;
 }
 
 const Alert = ({
@@ -19,6 +20,7 @@ const Alert = ({
   hasBorder = true,
   icon,
   variant = "info",
+  onClose,
   className,
 }: AlertProps) => {
   return (
@@ -29,7 +31,18 @@ const Alert = ({
       role="alert"
     >
       {icon && <AlertIcon variant={variant} name={icon} size={24} />}
+
       <div>{children}</div>
+
+      {onClose && (
+        <CloseIcon
+          variant="info"
+          name="close"
+          size={24}
+          onClick={onClose}
+          data-testid="alert-close-button"
+        />
+      )}
     </AlertRoot>
   );
 };
