+++ frontend/src/metabase/common/hooks/use-instance-locale/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-instance-locale";
