+++ frontend/src/metabase/common/hooks/use-instance-locale/use-instance-locale.ts
@@ -1,0 +1,6 @@
+import { useSelector } from "metabase/lib/redux";
+import { getSetting } from "metabase/selectors/settings";
+
+export const useInstanceLocale = () => {
+  return useSelector((state) => getSetting(state, "site-locale"));
+};
