+++ frontend/src/metabase/css/core/colors.module.css
@@ -2,6 +2,7 @@
  * NOTE: KEEP SYNCHRONIZED WITH:
  * frontend/src/metabase/ui/utils/colors.ts
  * frontend/src/metabase/lib/colors/palette.ts
+ * frontend/src/metabase/lib/colors/colors.ts
  * frontend/src/metabase/styled-components/containers/GlobalStyles/GlobalStyles.tsx
  * frontend/src/metabase/styled-components/theme/css-variables.ts
  */
