+++ frontend/src/metabase/embedding-sdk/theme/define-metabase-theme.ts
@@ -1,0 +1,4 @@
+import type { MetabaseTheme } from "./MetabaseTheme";
+
+export const defineMetabaseTheme = (theme: MetabaseTheme): MetabaseTheme =>
+  theme;
