+++ frontend/src/metabase/embedding-sdk/theme/index.ts
@@ -1,8 +1,5 @@
-import type { MetabaseTheme } from "./MetabaseTheme";
-
 export * from "./default-component-theme";
 // eslint-disable-next-line no-literal-metabase-strings -- file name
 export * from "./MetabaseTheme";
 
-export const defineMetabaseTheme = (theme: MetabaseTheme): MetabaseTheme =>
-  theme;
+export * from "./define-metabase-theme";
