+++ frontend/src/metabase/lib/colors/index.ts
@@ -1,1 +1,2 @@
 export * from "./palette";
+export * from "./colors";
