+++ frontend/src/metabase/static-viz/lib/colors.ts
@@ -1,4 +1,5 @@
-import { color, colors } from "metabase/lib/colors/palette";
+import { colors } from "metabase/lib/colors/colors";
+import { color } from "metabase/lib/colors/palette";
 import type { ColorPalette } from "metabase/lib/colors/types";
 import type { ColorGetter } from "metabase/visualizations/types";
